move validate payload for prediction into util for easier testing
